+++ halotools/mock_observables/two_point_clustering/tests/test_wp.py
@@ -31,8 +31,7 @@ def test_wp_auto_nonperiodic():
         randoms = np.random.random((Npts, 3))
 
     result = wp(sample1, rp_bins, pi_max, sample2=None,
-                randoms=randoms, period=None,
-                max_sample_size=int(1e4), estimator='Natural')
+                randoms=randoms, period=None, estimator='Natural')
 
     print(result)
     assert result.ndim == 1, "More than one correlation function returned erroneously."
@@ -47,8 +46,7 @@ def test_wp_auto_periodic():
         sample1 = np.random.random((Npts, 3))
 
     result = wp(sample1, rp_bins, pi_max, sample2=None,
-                randoms=None, period=period,
-                max_sample_size=int(1e4), estimator='Natural')
+                randoms=None, period=period, estimator='Natural')
 
     assert result.ndim == 1, "More than one correlation function returned erroneously."
 
@@ -63,8 +61,7 @@ def test_wp_cross_periodic():
         sample2 = np.random.random((Npts, 3))
 
     result = wp(sample1, rp_bins, pi_max, sample2=sample2,
-                randoms=None, period=period,
-                max_sample_size=int(1e4), estimator='Natural')
+                randoms=None, period=period, estimator='Natural')
 
     assert len(result) == 3, "wrong number of correlations returned"
     assert result[0].ndim == 1, "dimension of auto incorrect"
@@ -83,8 +80,7 @@ def test_wp_cross_nonperiodic():
         randoms = np.random.random((Npts, 3))
 
     result = wp(sample1, rp_bins, pi_max, sample2=sample2,
-                randoms=randoms, period=None,
-                max_sample_size=int(1e4), estimator='Natural')
+                randoms=randoms, period=None, estimator='Natural')
 
     assert len(result) == 3, "wrong number of correlations returned"
     assert result[0].ndim == 1, "dimension of auto incorrect"